# servers/api/swagger.yaml
openapi: 3.0.3
info:
  title: Python PvP Platform API
  description: |
    PvP coding platform — users write Python strategies that compete in
    head-to-head games. The simulator runs in an AWS Lambda, communicating
    with the API via a root token (LAMBDA_CALLBACK_TOKEN).

    ## Auth modes

    | Mode | Header | Used by |
    |------|--------|---------|
    | Public | None | Login only |
    | User | `Authorization: Bearer <session_id>` | User routes |
    | Root | `Authorization: Bearer <LAMBDA_CALLBACK_TOKEN>` | Admin routes, Lambda |

    Root tokens bypass all role checks and can call *any* route.
  version: 1.0.0
  contact:
    name: Python PvP Platform
servers:
  - url: http://localhost:3000
    description: Local development (via SSH tunnel to RDS)
paths:
...
  /competition:
    get:
      operationId: listCompetitions
      summary: List all competitions
      tags: [User]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: List of competitions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Competition'
        '500':
          $ref: '#/components/responses/InternalError'
  /competition/{id}:
    get:
      operationId: getCompetition
      summary: Get a competition by ID
      tags: [User]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Competition
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Competition'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /user/session:
    delete:
      operationId: logout
      summary: Logout — terminate current session
      tags: [User]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Session terminated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Session terminated
        '500':
          $ref: '#/components/responses/InternalError'
  /code:
    post:
      operationId: createCode
      summary: Create a new code with initial snapshot
      tags: [User]
      security: [{ BearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 50
                  example: my-bot-v1
                code:
                  type: string
                  description: Initial code text (optional — snapshot created only if provided)
                  example: "def update(game_states):\n    return {'alpha1': 0.5, 'alpha2': -0.3}"
      responses:
        '201':
          description: Code created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  name:
                    type: string
                  created_at_utc:
                    type: string
                    format: date-time
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      operationId: listMyCodes
      summary: List my codes with latest code and tested status
      tags: [User]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: List of codes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Code'
        '500':
          $ref: '#/components/responses/InternalError'
  /code/{code_id}:
    put:
      operationId: updateCode
      summary: Update code — creates a new snapshot
      tags: [User]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: code_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  example: "def update(game_states):\n    return {'alpha1': 1.0, 'alpha2': -1.0}"
      responses:
        '200':
          description: Code updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Code updated
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

    get:
      operationId: getCode
      summary: Get code details with latest code and tested status
      tags: [User]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: code_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Code details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Code'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /enroll:
    get:
      operationId: listMyEnrollments
      summary: List my enrollments
      tags: [User]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: List of enrollments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Enroll'
        '500':
          $ref: '#/components/responses/InternalError'
  /enroll/{enroll_id}:
    get:
      operationId: getEnrollment
      summary: Get enrollment details
      tags: [User]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: enroll_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Enrollment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Enroll'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /enroll/{enroll_id}/code:
    get:
      operationId: getLinkedCode
      summary: Get the code linked to an enrollment
      tags: [User]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: enroll_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Linked code (or null if none)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Code'
                  - nullable: true
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      operationId: linkCode
      summary: Link/replace code for an enrollment
      description: Deletes any existing link, then inserts the new one (at most one code per enrollment).
      tags: [User]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: enroll_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code_id]
              properties:
                code_id:
                  type: string
                  format: uuid
                  description: Must be owned by the current user
      responses:
        '200':
          description: Code linked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Code linked to enrollment
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalError'
  /enroll/{enroll_id}/code/{code_id}:
    delete:
      operationId: unlinkCode
      summary: Unlink code from enrollment
      tags: [User]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: enroll_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: code_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Code unlinked
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Code unlinked
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── Tests ───────────────────────────────────────────────────────────────
...
  /test:
    get:
      operationId: listAllMyTests
      summary: List all my test battles across all competitions
      tags: [User, Test]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: List of test battles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Battle'
        '500':
          $ref: '#/components/responses/InternalError'
  /test/{id}:
    get:
      operationId: getTest
      summary: Get a specific test battle result
      tags: [User, Test]
      security: [{ BearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Battle details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Battle'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  # ── Battles ─────────────────────────────────────────────────────────────
...
  /battle:
    get:
      operationId: listAllMyBattles
      summary: List all my battles across all competitions
      tags: [User, Battle]
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: List of battles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Battle'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |
        For user routes: session token from POST /public/user/session.
        For admin routes: LAMBDA_CALLBACK_TOKEN (long-lived root token).
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    Competition:
      type: object
      properties:
        id:
          type: string
          format: uuid
        display_name:
          type: string
        description:
          type: string
        start_time_utc:
          type: string
          format: date-time
        end_time_utc:
          type: string
          format: date-time
        game_reference:
          type: string
        helper_reference:
          type: string
        manifest_reference:
          type: string
        npc_user_id:
          type: string
          format: uuid
        created_at_utc:
          type: string
          format: date-time
        updated_at_utc:
          type: string
          format: date-time
...
  responses:
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'